Solutions to problem 19 and 20
Problem 19: The solution to problem 19 involves looping through all of the days since 01/01/1901 and checking for instances where a sunday occurs on the 1st of any month. No standard library date and time features are used.
Problem 20: The solution to problem 20 relies on a recycled but changed version of the algorithim used to solve problem 16. The solution is not particularly efficient and may be changed in the future.