Solution to problem 40
This solution uses an array containing the number of digits required to count through all the possible numbers of a given length. The digit at any location in the concatenated number can be determined by using the array to first find what range of values the number containing the digit lies in (i.e the length of the number), then simply use the remaing digits and the lower boundary for the range to find the number containing the digit, and the digit.