Solution to problem 41
Due the nature of this problem a non-brute force approach is not possible. This solution uses the property that a number is divisible by 3 if its digit sum is also divisible by 3 to significantly reduce the number of primes that must be considered. All 8 and 9 digit pandigital numbers are divisible by 3 and as a result the upper bound can be set to 7654321