Solution to problem 44
This solution uses the inverse function of the function used to calculate the terms to verify if a value is a term in the sequence.  As the difference between terms only increases (positive derivative) looping from term 1 and increasing guarantees that the first two terms to satisfy the criteria will have the smallest difference.